pyproject.toml
[tool.poetry]
name = "genesis-mobile-robot"
version = "0.1.0"
description = "Ambiente de simulação com Genesis para robô móvel tipo Pioneer com suporte a RL e planners clássicos"
authors = [
    "Rodrigo Santos <rodrigo.silva@dcc.ufmg.br>",
    "Bruno Lemos <bruno.lemos@dcc.ufmg.br>",
]
readme = "README.md"
license = "MIT"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
genesis-world = "==0.3.6"
torch = "==2.9.0"
numpy = "==2.2.6"
matplotlib = "==3.10.7"
scipy = "==1.16.3"
pandas = "==2.3.3"
networkx = "==3.5"
opencv-python = "==4.12.0.88"
tqdm = "==4.67.1"
pyyaml = "==6.0.3"
gymnasium = "^1.2.2"
stable-baselines3 = "^2.7.0"
tensorboard = "^2.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "==8.4.2"
pytest-cov = "==5.0.0"
ruff = "==0.14.4"
black = "==24.10.0"
isort = "==5.13.2"
pre-commit = "==4.4.0"
ipykernel = "==6.31.0"
tomli-w = "==1.2.0"

[tool.ruff]
line-length = 100
target-version = "py313"
extend-select = [
    "I",
    "UP",
    "N",
    "C4",
]
fix = true

[tool.black]
line-length = 100
target-version = [
    "py313",
]

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
